mark rfcomm_cid as 16 bit in rfcomm_send_internal - thanks to bob for spotting this

+++ trunk/src/rfcomm.h
@@ -59,7 +59,7 @@ void rfcomm_unregister_service_internal(uint8_t service_channel);
 void rfcomm_accept_connection_internal(uint16_t rfcomm_cid);
 void rfcomm_decline_connection_internal(uint16_t rfcomm_cid);
 void rfcomm_grant_credits(uint16_t rfcomm_cid, uint8_t credits);
-int  rfcomm_send_internal(uint8_t rfcomm_cid, uint8_t *data, uint16_t len);
+int  rfcomm_send_internal(uint16_t rfcomm_cid, uint8_t *data, uint16_t len);
 void rfcomm_close_connection(void *connection);
 
 #define UNLIMITED_INCOMING_CREDITS 0xff
